fix: address code review comments on new NPC/quest creation
